Move TaskProcessor-related types to linera_base
