+++ linera-base/src/lib.rs
@@ -34,6 +34,7 @@ pub mod port;
 pub mod prometheus_util;
 #[cfg(not(chain))]
 pub mod task;
+pub mod task_processor;
 pub mod time;
 #[cfg(test)]
 mod unit_tests;
